specify xesmf minimum version
There's a bug in an xesmf (I think 0.5.1) in which "longitude"/"latitude" is the only acceptable input format for the grid names. It's causing the test routine to fail on github. This issue doesn't come up when run offline - in which the test environment has the latest xesmf version (0.5.2). Not sure why the github test environment doesn't download the latest version...