Fix undercounting when pixels have nans, add a fix_ds call to aggregate
Previously, if pixels had nan values, these were not edited out of the weighting function - so the sum of the weights < 1, and the aggregated variable value would be lower than what it should be. The code can now deal with pixels where every datapoint is nan (in which case the weight vector is re-normalized to exclude that pixel(s)); and throws a warning if a pixel has a few (but not all) data points be nan (to be fully fixed in the future).

Other fixes include adding fix_ds(ds) to the beginning of aggregate; otherwise fix_ds(ds) would have to be called by the end user, which is not ideal.

cf_xarray is now provisionally listed as an explicit dependence (but *not yet* in the official testing environment) in an ill-placed attempt to fix the KeyError "lon" issue in testing, coming from cf_xarray in xesmf. Still unclear how to fix this - allegedly the new version of cf_xarray and xesmf have fixed this bug, but I can't get it out of this testing routine yet.